fix(frontend): surface credit grant visibility warnings
