+++ frontend/src/hooks/use-billing-credits.test.tsx
@@ -4,6 +4,7 @@ import type { PropsWithChildren } from "react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import {
   useActiveCreditGrants,
+  useAdminCreditGrants,
   useCurrentAllowances,
   useIssueCreditGrant,
 } from "./use-billing-credits";
@@ -29,12 +30,31 @@ function wrapperFactory() {
 beforeEach(() => vi.clearAllMocks());
 
 describe("billing credit hooks", () => {
+  it("loads a specific admin grant page without a silent fixed cutoff", async () => {
+    mockGet.mockResolvedValue({ grants: [], page: 2, per_page: 50, total: 75 });
+    const grants = renderHook(() => useAdminCreditGrants(2, 50), {
+      wrapper: wrapperFactory(),
+    });
+
+    await waitFor(() => expect(grants.result.current.isSuccess).toBe(true));
+    expect(mockGet).toHaveBeenCalledWith(
+      "/admin/credits/grants?page=2&per_page=50",
+    );
+  });
+
   it("normalizes all-owner and all-service grant payloads", async () => {
     mockPost.mockResolvedValue({
       batch_id: "batch-1",
       created_count: 3,
       activated_count: 3,
       pending_activation_count: 0,
+      recipients: [
+        {
+          recipient_user_id: "user-1",
+          recipient_billing_enabled: false,
+          activation_state: "active",
+        },
+      ],
     });
     const { result } = renderHook(() => useIssueCreditGrant(), {
       wrapper: wrapperFactory(),
@@ -85,9 +105,11 @@ describe("billing credit hooks", () => {
 
   it("scopes organization benefit reads without sharing personal cache paths", async () => {
     mockGet.mockImplementation((path: string) =>
-      Promise.resolve(path.includes("grants")
-        ? { grants: [], page: 1, per_page: 0, total: 0 }
-        : { allowances: [] }),
+      Promise.resolve(
+        path.includes("grants")
+          ? { grants: [], page: 1, per_page: 0, total: 0 }
+          : { allowances: [] },
+      ),
     );
     const grants = renderHook(() => useActiveCreditGrants("org/one"), {
       wrapper: wrapperFactory(),
@@ -98,9 +120,7 @@ describe("billing credit hooks", () => {
 
     await waitFor(() => expect(grants.result.current.isSuccess).toBe(true));
     await waitFor(() => expect(allowances.result.current.isSuccess).toBe(true));
-    expect(mockGet).toHaveBeenCalledWith(
-      "/billing/grants?owner_id=org%2Fone",
-    );
+    expect(mockGet).toHaveBeenCalledWith("/billing/grants?owner_id=org%2Fone");
     expect(mockGet).toHaveBeenCalledWith(
       "/billing/allowances?owner_id=org%2Fone",
     );
