+++ frontend/src/pages/billing.tsx
@@ -1,24 +1,13 @@
 import { Fragment, useMemo, useState } from "react";
 import { toast } from "sonner";
-import {
-  ChevronDown,
-  ChevronRight,
-  ExternalLink,
-  Info,
-  Plus,
-  WalletCards,
-  Coins,
-  Gauge,
-} from "lucide-react";
+import { ChevronRight } from "lucide-react";
 import { ApiError } from "@/lib/api-client";
 import { openExternal } from "@/lib/navigation";
-import { formatRelativeTime } from "@/lib/utils";
 import {
   BILLING_USAGE_PERIODS,
   type BillingMetric,
   type BillingUsagePeriod,
   type BillingUsageRow,
-  type BillingWalletResponse,
 } from "@/schemas/billing";
 import {
   useBillingUsage,
@@ -28,23 +17,11 @@ import {
   useTopUpHistory,
   openInvoiceReceipt,
 } from "@/hooks/use-billing";
-import {
-  useActiveCreditGrants,
-  useCurrentAllowances,
-} from "@/hooks/use-billing-credits";
-import { Button, ButtonIcon } from "@/components/ui/button";
+import { BillingBenefits } from "@/components/billing/billing-benefits";
+import { BillingWalletCard } from "@/components/billing/billing-wallet-card";
+import { Button } from "@/components/ui/button";
 import { Badge } from "@/components/ui/badge";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/components/ui/dialog";
-import { Input } from "@/components/ui/input";
 import {
   Select,
   SelectContent,
@@ -53,11 +30,6 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { Skeleton } from "@/components/ui/skeleton";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 import { ErrorBanner } from "@/components/shared/error-banner";
 import {
   Table,
@@ -67,9 +39,6 @@ import {
   TableHeader,
   TableRow,
 } from "@/components/ui/table";
-
-const DEFAULT_TOP_UP_CREDITS = 100;
-const TOP_UP_PRESETS = [100, 500, 1_000, 5_000] as const;
 
 export function BillingPage() {
   const [period, setPeriod] = useState<BillingUsagePeriod>("30d");
@@ -150,7 +119,7 @@ export function BillingPage() {
         </div>
       )}
 
-      <WalletCard
+      <BillingWalletCard
         wallet={wallet}
         loading={walletQuery.isLoading}
         unavailable={walletUnavailable}
@@ -173,118 +142,6 @@ export function BillingPage() {
 
       <TopUpHistory key={period} period={period} />
     </div>
-  );
-}
-
-function BillingBenefits() {
-  const grantsQuery = useActiveCreditGrants();
-  const allowancesQuery = useCurrentAllowances();
-  const grants = grantsQuery.data?.grants ?? [];
-  const allowances = allowancesQuery.data?.allowances ?? [];
-  const rolloutHidden = [grantsQuery.error, allowancesQuery.error].every(
-    (error) => error instanceof ApiError && error.status === 403,
-  );
-
-  if (rolloutHidden) return null;
-  if (grantsQuery.isLoading || allowancesQuery.isLoading) {
-    return <Skeleton className="h-40 w-full" />;
-  }
-  const visibleError = [grantsQuery.error, allowancesQuery.error].find(
-    (error) => error && !(error instanceof ApiError && error.status === 403),
-  );
-  if (visibleError) {
-    return (
-      <ErrorBanner
-        message={errorMessage(visibleError, "Failed to load billing credits")}
-        onRetry={() => {
-          void grantsQuery.refetch();
-          void allowancesQuery.refetch();
-        }}
-      />
-    );
-  }
-  if (grants.length === 0 && allowances.length === 0) return null;
-
-  return (
-    <Card>
-      <CardHeader className="flex-row items-center justify-between space-y-0">
-        <CardTitle>Included credits</CardTitle>
-        <Coins className="h-4 w-4 text-text-tertiary" />
-      </CardHeader>
-      <CardContent className="grid gap-6 md:grid-cols-2">
-        <section className="min-w-0">
-          <div className="mb-2 flex items-center gap-2 text-[12px] font-medium">
-            <Coins className="h-3.5 w-3.5 text-success" /> Credit grants
-          </div>
-          {grants.length === 0 ? (
-            <p className="text-[12px] text-muted-foreground">
-              No active grants.
-            </p>
-          ) : (
-            <div className="divide-y divide-border">
-              {grants.map((grant) => (
-                <div
-                  key={grant.id}
-                  className="flex items-start justify-between gap-3 py-2 first:pt-0"
-                >
-                  <div className="min-w-0">
-                    <p className="truncate text-[12px] font-medium">
-                      {grant.scope.all_services
-                        ? "All services"
-                        : grant.scope.service_slugs.join(", ")}
-                    </p>
-                    <p className="text-[11px] text-muted-foreground">
-                      {grant.expires_at
-                        ? `Expires ${new Date(grant.expires_at).toLocaleString()}`
-                        : "No expiry"}
-                    </p>
-                  </div>
-                  <span className="shrink-0 text-[12px] font-medium">
-                    {formatEstimatedCredits(grant.remaining_micros)}
-                  </span>
-                </div>
-              ))}
-            </div>
-          )}
-        </section>
-        <section className="min-w-0">
-          <div className="mb-2 flex items-center gap-2 text-[12px] font-medium">
-            <Gauge className="h-3.5 w-3.5 text-info" /> Free usage
-          </div>
-          {allowances.length === 0 ? (
-            <p className="text-[12px] text-muted-foreground">
-              No active allowances.
-            </p>
-          ) : (
-            <div className="divide-y divide-border">
-              {allowances.map(
-                ({ allowance, remaining_quantity, period_end }) => (
-                  <div
-                    key={allowance.id}
-                    className="flex items-start justify-between gap-3 py-2 first:pt-0"
-                  >
-                    <div className="min-w-0">
-                      <p className="truncate text-[12px] font-medium">
-                        {allowance.service_slug}
-                      </p>
-                      <p className="text-[11px] text-muted-foreground">
-                        {labelize(allowance.metric)}
-                        {period_end
-                          ? ` / resets ${new Date(period_end).toLocaleString()}`
-                          : " / one time"}
-                      </p>
-                    </div>
-                    <span className="shrink-0 text-[12px] font-medium">
-                      {formatNumber(remaining_quantity)} left
-                    </span>
-                  </div>
-                ),
-              )}
-            </div>
-          )}
-        </section>
-      </CardContent>
-    </Card>
   );
 }
 
@@ -450,351 +307,6 @@ function TopUpHistory({ period }: { readonly period: BillingUsagePeriod }) {
         )}
       </CardContent>
     </Card>
-  );
-}
-
-function WalletCard({
-  wallet,
-  loading,
-  unavailable,
-  error,
-  onRetry,
-  onProvision,
-  provisioning,
-  billingReady,
-  onTopUp,
-  topUpPending,
-}: {
-  readonly wallet: BillingWalletResponse | undefined;
-  readonly loading: boolean;
-  readonly unavailable: boolean;
-  readonly error: unknown;
-  readonly onRetry: () => void;
-  readonly onProvision: () => void;
-  readonly provisioning: boolean;
-  readonly billingReady: boolean;
-  readonly onTopUp: (amountCredits: number) => Promise<void>;
-  readonly topUpPending: boolean;
-}) {
-  const [showBreakdown, setShowBreakdown] = useState(false);
-
-  if (loading) {
-    return <Skeleton className="h-[170px] w-full" />;
-  }
-
-  if (!wallet && unavailable) {
-    return (
-      <Card>
-        <CardHeader className="flex-row items-center justify-between space-y-0">
-          <div>
-            <CardTitle>Wallet</CardTitle>
-            <p className="mt-1 text-[12px] text-muted-foreground">
-              No wallet provisioned.
-            </p>
-          </div>
-          <WalletCards className="h-4 w-4 text-text-tertiary" />
-        </CardHeader>
-        <CardContent>
-          <Button
-            variant="primary"
-            disabled={!billingReady || provisioning}
-            isLoading={provisioning}
-            onClick={onProvision}
-          >
-            <ButtonIcon variant="primary">
-              <Plus className="h-3 w-3" />
-            </ButtonIcon>
-            Provision Wallet
-          </Button>
-        </CardContent>
-      </Card>
-    );
-  }
-
-  if (!wallet && error) {
-    return (
-      <ErrorBanner
-        message={errorMessage(error, "Failed to load billing wallet.")}
-        onRetry={onRetry}
-      />
-    );
-  }
-
-  if (!wallet) {
-    return null;
-  }
-
-  return (
-    <Card>
-      <CardHeader className="flex-row items-center justify-between space-y-0">
-        <CardTitle>Wallet</CardTitle>
-        {wallet.suspended && <Badge variant="destructive">Suspended</Badge>}
-      </CardHeader>
-      <CardContent>
-        <div className="flex items-start justify-between gap-3">
-          <div className="min-w-0">
-            <div className="flex items-center gap-1.5">
-              <span className="text-[11px] text-muted-foreground">
-                Available
-              </span>
-              <Tooltip>
-                <TooltipTrigger asChild>
-                  <button
-                    type="button"
-                    aria-label="What available credits means"
-                    className="text-text-tertiary transition-colors hover:text-foreground"
-                  >
-                    <Info className="h-3 w-3" />
-                  </button>
-                </TooltipTrigger>
-                <TooltipContent
-                  side="top"
-                  className="max-w-[260px] leading-relaxed"
-                >
-                  Credits spendable now after in-flight requests, unsettled
-                  usage, and purchased-credit expiry holds. 1 credit is 1 USD.
-                </TooltipContent>
-              </Tooltip>
-            </div>
-            <div className="mt-1 truncate text-[28px] font-semibold leading-tight">
-              {formatCredits(wallet.available_credits)}
-            </div>
-            <div className="mt-1 text-[11px] text-muted-foreground">
-              Updated {formatRelativeTime(wallet.balance_synced_at)}
-            </div>
-            <button
-              type="button"
-              aria-expanded={showBreakdown}
-              aria-controls="wallet-breakdown"
-              className="mt-2 flex items-center gap-1 text-[11px] text-muted-foreground transition-colors hover:text-foreground"
-              onClick={() => setShowBreakdown((current) => !current)}
-            >
-              {showBreakdown ? "Hide breakdown" : "View breakdown"}
-              <ChevronDown
-                className={`h-3 w-3 transition-transform ${
-                  showBreakdown ? "rotate-180" : ""
-                }`}
-              />
-            </button>
-          </div>
-          <AddCreditsDialog
-            billingReady={billingReady}
-            pending={topUpPending}
-            onTopUp={onTopUp}
-          />
-        </div>
-
-        {showBreakdown && (
-          <div id="wallet-breakdown">
-            <hr className="my-4 border-border" />
-            <p className="text-[11px] leading-relaxed text-muted-foreground">
-              Not all of your balance is spendable at any moment — requests in
-              flight hold credits until they finish.
-            </p>
-            <div className="mt-3 space-y-1.5">
-              <BreakdownRow
-                label="Balance"
-                hint="Last provider-synced balance"
-                value={formatCredits(wallet.balance_credits)}
-              />
-              <BreakdownRow
-                label="Reserved"
-                hint="Held for requests in flight"
-                value={formatCredits(wallet.reserved_credits)}
-              />
-              <BreakdownRow
-                label="Pending"
-                hint="Charged, awaiting provider sync"
-                value={formatCredits(wallet.pending_lago_debits)}
-              />
-              {wallet.pending_topup_expiry_credits > 0 && (
-                <BreakdownRow
-                  label="Expiring"
-                  hint="Held while expired purchases are removed"
-                  value={formatCredits(wallet.pending_topup_expiry_credits)}
-                />
-              )}
-              {wallet.overdraft_cap_credits > 0 && (
-                <BreakdownRow
-                  label="Overdraft"
-                  hint="Extra capacity beyond your balance"
-                  value={formatCredits(wallet.overdraft_cap_credits)}
-                />
-              )}
-            </div>
-            <hr className="my-3 border-border" />
-            <div className="flex items-center justify-between text-[11px] text-muted-foreground">
-              <span>Plan</span>
-              <span className="text-foreground">
-                {labelize(wallet.plan_kind)}
-              </span>
-            </div>
-            <p className="mt-2 text-[11px] text-muted-foreground">
-              Available = Balance - Reserved - Pending - Expiring.
-            </p>
-          </div>
-        )}
-      </CardContent>
-    </Card>
-  );
-}
-
-/**
- * Adding credits is the wallet's primary verb, so it lives on the wallet card.
- * The amount form is bulky enough to crowd out the balance, so it opens in a
- * dialog rather than sitting inline.
- */
-function AddCreditsDialog({
-  billingReady,
-  pending,
-  onTopUp,
-}: {
-  readonly billingReady: boolean;
-  readonly pending: boolean;
-  readonly onTopUp: (amountCredits: number) => Promise<void>;
-}) {
-  const [open, setOpen] = useState(false);
-  const [credits, setCredits] = useState(String(DEFAULT_TOP_UP_CREDITS));
-
-  const amount = Number(credits);
-  const amountValid =
-    Number.isInteger(amount) && amount > 0 && amount <= 10_000_000;
-  const submitDisabled = !billingReady || !amountValid || pending;
-
-  const trigger = (
-    <Button variant="primary" disabled={!billingReady || pending}>
-      <ButtonIcon variant="primary">
-        <Plus className="h-3 w-3" />
-      </ButtonIcon>
-      Add credits
-    </Button>
-  );
-
-  // A disabled trigger swallows pointer events, so the reason has to sit on a
-  // wrapper rather than the button itself.
-  if (!billingReady) {
-    return (
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <span tabIndex={0}>{trigger}</span>
-        </TooltipTrigger>
-        <TooltipContent side="left">
-          Payments aren&apos;t enabled on this deployment yet.
-        </TooltipContent>
-      </Tooltip>
-    );
-  }
-
-  return (
-    <Dialog open={open} onOpenChange={setOpen}>
-      <DialogTrigger asChild>{trigger}</DialogTrigger>
-      <DialogContent className="sm:max-w-[420px]">
-        <DialogHeader>
-          <DialogTitle>Add credits</DialogTitle>
-          <DialogDescription>
-            1 credit = 1 USD. Purchased credits expire one year after payment.
-          </DialogDescription>
-        </DialogHeader>
-
-        <div className="space-y-4">
-          <div className="space-y-2">
-            <span className="text-[12px] font-medium">Amount</span>
-            <div className="flex flex-wrap gap-2">
-              {TOP_UP_PRESETS.map((preset) => (
-                <Button
-                  key={preset}
-                  variant={amount === preset ? "secondary" : "outline"}
-                  size="sm"
-                  disabled={pending}
-                  onClick={() => setCredits(String(preset))}
-                >
-                  {formatNumber(preset)}
-                </Button>
-              ))}
-            </div>
-          </div>
-
-          <div className="space-y-2">
-            <label className="text-[12px] font-medium" htmlFor="topup-credits">
-              Or enter an amount
-            </label>
-            <div className="flex items-center gap-3">
-              <Input
-                id="topup-credits"
-                type="number"
-                min={1}
-                max={10_000_000}
-                step={1}
-                value={credits}
-                onChange={(event) => setCredits(event.target.value)}
-                disabled={pending}
-                aria-describedby="topup-total"
-              />
-              <span
-                id="topup-total"
-                className="shrink-0 text-[13px] text-muted-foreground"
-              >
-                {formatUsd(amount)}
-              </span>
-            </div>
-          </div>
-
-          <p className="text-[11px] leading-relaxed text-muted-foreground">
-            We&apos;ll hand you to Stripe to pay, then bring you back here.
-            Credits land once the payment clears.
-          </p>
-        </div>
-
-        <DialogFooter>
-          <Button variant="outline" onClick={() => setOpen(false)}>
-            Cancel
-          </Button>
-          <Button
-            variant="primary"
-            disabled={submitDisabled}
-            isLoading={pending}
-            onClick={() => void onTopUp(amount)}
-          >
-            <ButtonIcon variant="primary">
-              <ExternalLink className="h-3 w-3" />
-            </ButtonIcon>
-            Continue to payment
-          </Button>
-        </DialogFooter>
-      </DialogContent>
-    </Dialog>
-  );
-}
-
-function BreakdownRow({
-  label,
-  hint,
-  value,
-  emphasis = false,
-}: {
-  readonly label: string;
-  readonly hint: string;
-  readonly value: string;
-  readonly emphasis?: boolean;
-}) {
-  return (
-    <div className="flex items-baseline justify-between gap-3">
-      <div className="min-w-0">
-        <div className={emphasis ? "text-[12px] font-medium" : "text-[12px]"}>
-          {label}
-        </div>
-        <div className="text-[11px] text-muted-foreground">{hint}</div>
-      </div>
-      <div
-        className={
-          emphasis
-            ? "shrink-0 text-[13px] font-semibold"
-            : "shrink-0 text-[13px] text-muted-foreground"
-        }
-      >
-        {value}
-      </div>
-    </div>
   );
 }
 
@@ -1136,17 +648,6 @@ function labelize(value: string): string {
 
 function formatCredits(value: number): string {
   return `${formatNumber(value)} credits`;
-}
-
-/** Credits are 1:1 with USD, so the charge is worth stating outright. */
-function formatUsd(credits: number): string {
-  if (!Number.isFinite(credits) || credits <= 0) {
-    return "—";
-  }
-  return new Intl.NumberFormat(undefined, {
-    style: "currency",
-    currency: "USD",
-  }).format(credits);
 }
 
 function formatNumber(value: number): string {
